+++ ktrain/utils.py
@@ -71,7 +71,7 @@ def is_huggingface_from_model(model):
 
 
 def is_huggingface_from_data(data):
-    return type(data).__name__ == 'TransformerSequence'
+    return type(data).__name__ in ['TransformerDataset']
 
 
 
@@ -158,11 +158,6 @@ def shape_from_data(data):
     err_msg = 'could not determine shape from %s' % (type(data))
     if is_iter(data):
         if isinstance(data, Dataset): return data.xshape()
-        elif is_ner(data=data): return (len(data.x), data[0][0][0].shape[1])  # NERSequence
-        elif is_huggingface(data=data):  # HF Transformer
-            return (len(data.x), data[0][0][0].shape[1])
-        elif is_nodeclass(data=data):                                 # NodeSequence
-            return data[0][0][0].shape[1:]  # returns 1st neighborhood only
         elif hasattr(data, 'image_shape'): return data.image_shape          # DirectoryIterator/DataFrameIterator
         elif hasattr(data, 'x'):                                            # NumpyIterator
             return data.x.shape[1:]
@@ -185,8 +180,7 @@ def ondisk(data):
     if hasattr(data, 'ondisk'): return data.ondisk()
 
     ondisk = is_iter(data) and \
-             (type(data).__name__ not in  ['ArrayDataset', 'NumpyArrayIterator', 'NERSequence',
-                                           'NodeSequenceWrapper', 'TransformerSequence'])
+             (type(data).__name__ not in  ['NumpyArrayIterator'])
     return ondisk
 
 
@@ -194,14 +188,9 @@ def nsamples_from_data(data):
     err_msg = 'could not determine number of samples from %s' % (type(data))
     if is_iter(data):
         if isinstance(data, Dataset): return data.nsamples()
-        elif is_ner(data=data): return len(data.x)      # NERSequence
-        elif is_huggingface(data=data):  # HuggingFace Transformer
-            return len(data.x)
-        elif is_nodeclass(data=data):           # NodeSequenceWrapper
-            return data.targets.shape[0]
-        elif hasattr(data, 'samples'):                # DirectoryIterator/DataFrameIterator
+        elif hasattr(data, 'samples'):  # DirectoryIterator/DataFrameIterator
             return data.samples
-        elif hasattr(data, 'n'):                      # DirectoryIterator/DataFrameIterator/NumpyIterator
+        elif hasattr(data, 'n'):     # DirectoryIterator/DataFrameIterator/NumpyIterator
             return data.n
         else:
             raise Exception(err_msg)
@@ -218,16 +207,11 @@ def nsamples_from_data(data):
 def nclasses_from_data(data):
     if is_iter(data):
         if isinstance(data, Dataset): return data.nclasses()
-        elif is_ner(data=data): return len(data.p._label_vocab._id2token)    # NERSequence
-        elif is_huggingface(data=data):         # Hugging Face Transformer
-            return data.y.shape[1]
-        elif is_nodeclass(data=data):                                # NodeSequenceWrapper
-            return data[0][1].shape[1]                                   
-        elif hasattr(data, 'classes'):                                     # DirectoryIterator
+        elif hasattr(data, 'classes'):   # DirectoryIterator
             return len(set(data.classes))
         else:
             try:
-                return data[0][1].shape[1]                                 # DataFrameIterator/NumpyIterator
+                return data[0][1].shape[1]  # DataFrameIterator/NumpyIterator
             except:
                 raise Exception('could not determine number of classes from %s' % (type(data)))
     else:
@@ -240,11 +224,6 @@ def nclasses_from_data(data):
 def y_from_data(data):
     if is_iter(data):
         if isinstance(data, Dataset): return data.get_y()
-        elif is_ner(data=data): return data.y    # NERSequence
-        if is_huggingface(data=data):  # Hugging Face Transformer
-            return data.y
-        elif is_nodeclass(data=data):      # NodeSequenceWrapper
-            return data.targets
         elif hasattr(data, 'classes'): # DirectoryIterator
             return to_categorical(data.classes)
         elif hasattr(data, 'labels'):  # DataFrameIterator
@@ -263,8 +242,7 @@ def y_from_data(data):
 def is_iter(data, ignore=False):
     if ignore: return True
     iter_classes = ["NumpyArrayIterator", "DirectoryIterator",
-                    "DataFrameIterator", "Iterator", "Sequence", 
-                    "NERSequence", "NodeSequenceWrapper", "TransformerSequence"]
+                    "DataFrameIterator", "Iterator", "Sequence"]
     return data.__class__.__name__ in iter_classes or isinstance(data, Dataset)